Split StatsdUnitTests
