+++ tests/StatsdClient.Tests/MetricSerializerTests.cs
@@ -6,7 +6,7 @@ using NUnit.Framework;
 namespace StatsdClient.Tests
 {
     [TestFixture]
-    public class StatsdUnitTests
+    public class MetricSerializerTests
     {
         private IStatsdUDP _udp;
         private IRandomGenerator _randomGenerator;
@@ -512,130 +512,6 @@ namespace StatsdClient.Tests
             Mock.Get(_udp).Verify(x => x.Send("timer:1.1|ms"));
         }
 
-        // =-=-=-=- EVENT -=-=-=-=
-        // Event(string title, string text, string alertType = null, string aggregationKey = null, string sourceType = null, int? dateHappened = null, string priority = null, string hostname = null, string[] tags = null)
-
-        [Test]
-        public void Send_event()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("title", "text");
-            Mock.Get(_udp).Verify(x => x.Send("_e{5,4}:title|text"));
-        }
-
-        [Test]
-        public void Send_event_with_alertType()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("title", "text", alertType: "warning");
-            Mock.Get(_udp).Verify(x => x.Send("_e{5,4}:title|text|t:warning"));
-        }
-
-        [Test]
-        public void Send_event_with_aggregationKey()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("title", "text", aggregationKey: "key");
-            Mock.Get(_udp).Verify(x => x.Send("_e{5,4}:title|text|k:key"));
-        }
-
-        [Test]
-        public void Send_event_with_sourceType()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("title", "text", sourceType: "source");
-            Mock.Get(_udp).Verify(x => x.Send("_e{5,4}:title|text|s:source"));
-        }
-
-        [Test]
-        public void Send_event_with_dateHappened()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("title", "text", dateHappened: 123456);
-            Mock.Get(_udp).Verify(x => x.Send("_e{5,4}:title|text|d:123456"));
-        }
-
-        [Test]
-        public void Send_event_with_priority()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("title", "text", priority: "low");
-            Mock.Get(_udp).Verify(x => x.Send("_e{5,4}:title|text|p:low"));
-        }
-
-        [Test]
-        public void Send_event_with_hostname()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("title", "text", hostname: "hostname");
-            Mock.Get(_udp).Verify(x => x.Send("_e{5,4}:title|text|h:hostname"));
-        }
-
-        [Test]
-        public void Send_event_with_tags()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("title", "text", tags: new[] { "tag1", "tag2" });
-            Mock.Get(_udp).Verify(x => x.Send("_e{5,4}:title|text|#tag1,tag2"));
-        }
-
-        [Test]
-        public void Send_event_with_message_that_is_too_long()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-
-            var length = (8 * 1024) - 16; // 16 is the number of characters in the final message that is not the title
-            var builder = BuildLongString(length);
-            var title = builder;
-
-            var exception = Assert.Throws<Exception>(() => s.Send(title + "x", "text"));
-            Assert.That(exception.Message, Contains.Substring("payload is too big"));
-        }
-
-        [Test]
-        public void Send_event_with_truncation_for_title_that_is_too_long()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-
-            var length = (8 * 1024) - 16; // 16 is the number of characters in the final message that is not the title
-            var builder = BuildLongString(length);
-            var title = builder;
-
-            s.Send(title + "x", "text", truncateIfTooLong: true);
-            var expected = string.Format("_e{{{0},4}}:{1}|text", length, title);
-            Mock.Get(_udp).Verify(x => x.Send(expected));
-        }
-
-        [Test]
-        public void Send_event_with_truncation_for_text_that_is_too_long()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-
-            var length = (8 * 1024) - 17; // 17 is the number of characters in the final message that is not the text
-            var builder = BuildLongString(length);
-            var text = builder;
-
-            s.Send("title", text + "x", truncateIfTooLong: true);
-            var expected = string.Format("_e{{5,{0}}}:title|{1}", length, text);
-            Mock.Get(_udp).Verify(x => x.Send(expected));
-        }
-
-        [Test]
-        public void Send_event_with_statsd_truncation()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            // Enable truncation at Statsd level
-            s.TruncateIfTooLong = true;
-
-            var length = (8 * 1024) - 17; // 17 is the number of characters in the final message that is not the text
-            var builder = BuildLongString(length);
-            var text = builder;
-
-            s.Send("title", text + "x");
-            var expected = string.Format("_e{{5,{0}}}:title|{1}", length, text);
-            Mock.Get(_udp).Verify(x => x.Send(expected));
-        }
-
         // =-=-=-=- PREFIX -=-=-=-=
 
         [Test]
@@ -951,238 +827,6 @@ namespace StatsdClient.Tests
 
             Assert.That(s.Commands.Count, Is.EqualTo(1));
             Assert.That(s.Commands[0], Is.EqualTo("set:string|s|@0.5|#tag1:true,tag2"));
-        }
-
-        // =-=-=-=- ServiceCheck -=-=-=-=
-        [Test]
-        public void Send_service_check()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("name", 0);
-            Mock.Get(_udp).Verify(x => x.Send("_sc|name|0"));
-        }
-
-        [Test]
-        public void Send_service_check_with_timestamp()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("name", 0, timestamp: 1);
-            Mock.Get(_udp).Verify(x => x.Send("_sc|name|0|d:1"));
-        }
-
-        [Test]
-        public void Send_service_check_with_hostname()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("name", 0, hostname: "hostname");
-            Mock.Get(_udp).Verify(x => x.Send("_sc|name|0|h:hostname"));
-        }
-
-        [Test]
-        public void Send_service_check_with_tags()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("name", 0, tags: new[] { "tag1:value1", "tag2", "tag3:value3" });
-            Mock.Get(_udp).Verify(x => x.Send("_sc|name|0|#tag1:value1,tag2,tag3:value3"));
-        }
-
-        [Test]
-        public void Send_service_check_with_message()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("name", 0, serviceCheckMessage: "message");
-            Mock.Get(_udp).Verify(x => x.Send("_sc|name|0|m:message"));
-        }
-
-        [Test]
-        public void Send_service_check_with_pipe_in_name()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-
-            Assert.Throws<ArgumentException>(() => s.Send("name|", 0));
-        }
-
-        [Test]
-        [TestCase("\r\n")]
-        [TestCase("\n")]
-        public void Send_service_check_with_new_line_in_name(string newline)
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("name" + newline, 0);
-            Mock.Get(_udp).Verify(x => x.Send("_sc|name\\n|0"));
-        }
-
-        [Test]
-        public void Send_service_check_with_suffix_in_message()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("name", 0, serviceCheckMessage: "m:message");
-            Mock.Get(_udp).Verify(x => x.Send("_sc|name|0|m:m\\:message"));
-        }
-
-        [Test]
-        public void Send_service_check_with_all_optional()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Send("name", 0, 1, "hostname", new[] { "tag1:value1", "tag2", "tag3:value3" }, "message");
-            Mock.Get(_udp).Verify(x => x.Send("_sc|name|0|d:1|h:hostname|#tag1:value1,tag2,tag3:value3|m:message"));
-        }
-
-        [Test]
-        public void Send_service_check_with_message_that_is_too_long()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-
-            var length = (8 * 1024) - 13;
-            var builder = BuildLongString(length);
-            var message = builder;
-
-            var exception = Assert.Throws<Exception>(() => s.Send("name", 0, serviceCheckMessage: message + "x"));
-            Assert.That(exception.Message, Contains.Substring("payload is too big"));
-        }
-
-        [Test]
-        public void Send_service_check_with_message_that_is_too_long_truncate()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-
-            var length = (8 * 1024) - 13;
-            var builder = BuildLongString(length);
-            var message = builder;
-
-            s.Send("name", 0, serviceCheckMessage: message + "x", truncateIfTooLong: true);
-
-            var expected = "_sc|name|0|m:" + message;
-            Mock.Get(_udp).Verify(x => x.Send(expected));
-        }
-
-        [Test]
-        public void Send_service_check_with_name_that_is_too_long_truncate()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-
-            var length = (8 * 1024) - 6;
-            var builder = BuildLongString(length);
-            var name = builder;
-
-            var exception = Assert.Throws<ArgumentException>(() => s.Send(name + "x", 0, truncateIfTooLong: true));
-            Assert.That(exception.Message, Contains.Substring("payload is too big"));
-        }
-
-        [Test]
-        public void Add_service_check()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Add("name", 0);
-
-            Assert.That(s.Commands.Count, Is.EqualTo(1));
-            Assert.That(s.Commands[0], Is.EqualTo("_sc|name|0"));
-        }
-
-        [Test]
-        public void Add_service_check_with_timestamp()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Add("name", 0, timestamp: 1);
-
-            Assert.That(s.Commands.Count, Is.EqualTo(1));
-            Assert.That(s.Commands[0], Is.EqualTo("_sc|name|0|d:1"));
-        }
-
-        [Test]
-        public void Add_service_check_with_hostname()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Add("name", 0, hostname: "hostname");
-
-            Assert.That(s.Commands.Count, Is.EqualTo(1));
-            Assert.That(s.Commands[0], Is.EqualTo("_sc|name|0|h:hostname"));
-        }
-
-        [Test]
-        public void Add_service_check_with_tags()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Add("name", 0, tags: new[] { "tag1:value1", "tag2", "tag3:value3" });
-
-            Assert.That(s.Commands.Count, Is.EqualTo(1));
-            Assert.That(s.Commands[0], Is.EqualTo("_sc|name|0|#tag1:value1,tag2,tag3:value3"));
-        }
-
-        [Test]
-        public void Add_service_check_with_message()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Add("name", 0, serviceCheckMessage: "message");
-
-            Assert.That(s.Commands.Count, Is.EqualTo(1));
-            Assert.That(s.Commands[0], Is.EqualTo("_sc|name|0|m:message"));
-        }
-
-        [Test]
-        public void Add_service_check_with_pipe_in_name()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-
-            Assert.Throws<ArgumentException>(() => s.Add("name|", 0));
-        }
-
-        [Test]
-        [TestCase("\r\n")]
-        [TestCase("\n")]
-        public void Add_service_check_with_new_line_in_name(string newline)
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Add("name" + newline, 0);
-
-            Assert.That(s.Commands.Count, Is.EqualTo(1));
-            Assert.That(s.Commands[0], Is.EqualTo("_sc|name\\n|0"));
-        }
-
-        [Test]
-        public void Add_service_check_with_suffix_in_message()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Add("name", 0, serviceCheckMessage: "m:message");
-
-            Assert.That(s.Commands.Count, Is.EqualTo(1));
-            Assert.That(s.Commands[0], Is.EqualTo("_sc|name|0|m:m\\:message"));
-        }
-
-        [Test]
-        public void Add_service_check_with_all_optional()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-            s.Add("name", 0, 1, "hostname", new[] { "tag1:value1", "tag2", "tag3:value3" }, "message");
-
-            Assert.That(s.Commands.Count, Is.EqualTo(1));
-            Assert.That(s.Commands[0], Is.EqualTo("_sc|name|0|d:1|h:hostname|#tag1:value1,tag2,tag3:value3|m:message"));
-        }
-
-        [Test]
-        public void Add_service_check_with_message_that_is_too_long()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-
-            var length = (8 * 1024) - 13;
-            var builder = BuildLongString(length);
-            var message = builder;
-
-            var exception = Assert.Throws<Exception>(() => s.Add("name", 0, serviceCheckMessage: message + "x"));
-            Assert.That(exception.Message, Contains.Substring("payload is too big"));
-        }
-
-        [Test]
-        public void Add_service_check_with_name_that_is_too_long()
-        {
-            Statsd s = new Statsd(_udp, _randomGenerator, _stopwatch);
-
-            var length = (8 * 1024) - 6;
-            var builder = BuildLongString(length);
-            var name = builder;
-
-            var exception = Assert.Throws<Exception>(() => s.Add(name + "x", 0));
-            Assert.That(exception.Message, Contains.Substring("payload is too big"));
         }
 
         private static string BuildLongString(int length)
